Fix UnmanagedCallersOnlyTest by making the negative case a separate test
